Update authenticate middleware
I was looking through the code and saw that the authenticate wasn't actually returning the decoded token. This new code, according to the docs is the best way to return the decoded token while also verifying the token.  I noticed some weird behavior when trying to make a put request to /users. Take a look at it and make sure, but hopefully this will help.